+++ config/checkstyle/checkstyle.xml
@@ -1,0 +1,78 @@
+<?xml version="1.0"?>
+<!DOCTYPE module PUBLIC
+    "-//Checkstyle//DTD Checkstyle Configuration 1.3//EN"
+    "https://checkstyle.org/dtds/configuration_1_3.dtd">
+<module name="Checker">
+  <property name="severity" value="warning"/>
+
+  <!-- Line length -->
+  <module name="LineLength">
+    <property name="max" value="120"/>
+    <property name="ignorePattern" value="^package |^import |http[s]?://"/>
+  </module>
+
+  <!-- Import rules -->
+  <module name="AvoidStarImport"/>
+  <module name="UnusedImports"/>
+  <module name="ImportOrder">
+    <property name="groups" value="java,javax,org,com,rat"/>
+    <property name="ordered" value="true"/>
+    <property name="separateLineBetweenGroups" value="true"/>
+    <property name="sortStaticImportsAlphabetically" value="true"/>
+    <property name="sortImportsInGroupAlphabetically" value="true"/>
+  </module>
+
+  <module name="TreeWalker">
+    <!-- Naming -->
+    <module name="PackageName">
+      <property name="format" value="^[a-z]+(\.[a-z][a-z0-9]*)*$"/>
+    </module>
+    <module name="TypeName"/>
+    <module name="MethodName">
+      <property name="format" value="^[a-z][a-zA-Z0-9]*$"/>
+    </module>
+    <module name="ParameterName"/>
+    <module name="LocalVariableName"/>
+    <module name="MemberName"/>
+    <module name="ConstantName"/>
+
+    <!-- Braces and indentation -->
+    <module name="LeftCurly">
+      <property name="option" value="eol"/>
+    </module>
+    <module name="RightCurly">
+      <property name="option" value="alone_or_singleline"/>
+    </module>
+    <module name="NeedBraces"/>
+    <module name="Indentation">
+      <property name="basicOffset" value="4"/>
+      <property name="braceAdjustment" value="0"/>
+      <property name="caseIndent" value="0"/>
+      <property name="lineWrappingIndentation" value="8"/>
+    </module>
+
+    <!-- Whitespace -->
+    <module name="WhitespaceAround"/>
+    <module name="WhitespaceAfter"/>
+    <module name="NoWhitespaceBefore"/>
+    <module name="GenericWhitespace"/>
+
+    <!-- Switch fallthrough must be explicit -->
+    <module name="FallThrough">
+      <property name="reliefPattern" value="falls? through"/>
+    </module>
+
+    <!-- Array style -->
+    <module name="ArrayTypeStyle">
+      <property name="style" value="attached"/>
+    </module>
+
+  </module>
+
+  <!-- Suppress method naming rule in tests to allow underscores if desired -->
+  <module name="SuppressionFilter">
+    <property name="file" value="config/checkstyle/suppressions.xml"/>
+  </module>
+
+</module>
+
