+++ config/checkstyle/suppressions.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<!DOCTYPE suppressions PUBLIC
+    "-//Checkstyle//DTD SuppressionFilter Configuration 1.2//EN"
+    "https://checkstyle.org/dtds/suppressions_1_2.dtd">
+<suppressions>
+  <!-- Allow underscores in test method names if needed -->
+  <suppress files="src/test/java/.*" checks="MethodName"/>
+</suppressions>
+
