+++ src/main/java/rat/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package rat;
 
 import java.time.LocalDate;
