+++ src/main/java/rat/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package rat;
 
 import java.io.File;
 import java.io.FileWriter;
