+++ src/main/java/rat/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package rat;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
