+++ src/main/java/rat/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package rat;
 
 import java.util.Scanner;
