+++ text-ui-test/runtest.bat
@@ -13,7 +13,7 @@ if not exist data mkdir data
 if exist data\Rat.txt del data\Rat.txt
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\duke\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\rat\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -21,7 +21,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin duke.Rat < input.txt > ACTUAL.TXT
+java -classpath ..\bin rat.Rat < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
