+++ text-ui-test/runtest.sh
@@ -22,14 +22,14 @@ mkdir -p data
 rm -f data/Rat.txt
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/duke/*.java
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/rat/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin duke.Rat < input.txt > ACTUAL.TXT
+java -classpath ../bin rat.Rat < input.txt > ACTUAL.TXT
 
 # compare the output to the expected output
 diff ACTUAL.TXT EXPECTED.TXT
